+++ static-assets/components/cstudio-forms/forms-engine.js
@@ -1138,8 +1138,12 @@ const initializeCStudioForms = () => {
 							});
 					};
 
-					CrafterCMSNext.system.getStore().subscribe(() => {
+					CrafterCMSNext.system.getStore().subscribe((store) => {
 						getInitialConfiguration();
+
+						if (!store.getState().contentTypes?.byId) {
+							store.dispatch({ type: 'FETCH_CONTENT_TYPES' });
+						}
 					});
 				},